Bug 553020 - Add unforgeable Jetpack ID, r=atul,myk
This adds ECDSA (NIST256p) keypair generation, and code in "cfx xpi" that
creates a keypair when one does not already exist. The "jetpack id" (a hash
of the pubkey) is put in package.json . Eventually, the private key will be
used to sign the XPI, and the public key can be used to verify it.