Bug 591525: check require() paths against manifest generated at link-time
The general idea is to put expected linker behavior in the machine-readable
manifest, so that code-review tools can show a (human) reviewer what each
require("foo") statement is expected to load, and then the runtime code will
make sure that the expectations are met. This is intended to protect against
an attack in which the XPI is modified, after the manifest is generated, to
include a new module that the runtime linker will load in lieu of a module
that the manifest specified, thus violating the reviewer's expectations.

The manifest is now indexed by the resource: URL of the module being scanned,
and is a regular dictionary/object rather than a hard-to-scan list. It also
records more information about each module, including a hash of its contents,
which can eventually be checked by the runtime linker.