bootstrap-remote-process.js now registers a receiver for an 'addInjectedSandboxScript' event, which allows the parent process to define scripts to inject into any new sandboxes made by the require() implementation. This allowed me to remove all the confusing, hacky es5 and CommonJS-compliance related code from this file.
The 'cuddlefish' module now exports a symbol called 'es5code', which is the filename/contents info for es5.js, usable by any other module for injecting into desired sandboxes. Not a huge fan of this, but it seems like the simplest solution right now.

Removed duplicate Iterator-overriding logic from securable-module.js and bootstrap-remote-process.js, moving it into es5.js. It may not technically be es5, but it's part of the same blob of always-injected code. One alternative to this is to create a new file called better-iterator.js and inject it along with es5.js into all new sandboxes.