Bug 636693: 'cfx docs' should tolerate non-.js files in package dirs. r=myk
Editor tempfiles or other non *.js files were being recursed into as if they
were directories, causing a crash. The fix was to change the
packaging.build_pkg_index() data structure to distinguish between files and
directories, and to update webdocs.py (the only consumer) to match.

webdocs.py was also updated to tolerate lack of 'keywords' in package.json