Bug 647126 - dump resultsfile contents when unexpected. r=myk
A recent run of the fedora64 buildslave stopped just after doing some JS
tests, in a way that suggests the parent runner.py code never saw the "OK" or
"FAIL" results that it was expecting. This patch prints the contents of the
resultsfile when it contains something else, just in case the test is adding
some extra text or whitespace or something.