Bug 627607 - Rewrite manifest-generation code and runtime module loader.
For normal modules (i.e. user-written code and the SDK-provided modules
that it uses), the require() statement is now entirely driven by the
contents of the manifest. This allows reviewers to statically determine
what code will be used to build the module graph, and moves control of
the module search process out of the runtime code and into the linker.
Initial loader code and unit tests do not use the manifest, and instead
do a runtime search. The runtime code no longer does a "security policy"
lookup: instead the presence of a module name on the manifest indicates
that permission is granted, and a missing name causes an error (not a
mere warning).

The new manifest contains statements like "when module X does a
require(Y), it will get code Z", where X and Z are URIs (starting with
resource://JID), and Y is a module name (like "panel" or
"context-menu"). The manifest is a dictionary mapping from X to an
object full of of information about module X, including a
".requirements" property, which is a dictionary mapping from Y to an
object of "requirement data". For normal modules, the main property in
reqdata is "uri", which contains Z. For magic modules like "chrome" or
"self", reqdata may contain other information (like the base URI for
that package's data, allowing self.data.url() to work).

The main changes are in python-lib/cuddlefish/manifest.py (both the
linker-side module search logic and the manifest generation code), and
in packages/api-utils/lib/securable-module.js (runtime module-loading
code), followed by packages/api-utils/lib/cuddlefish.js (removing the
"security policy" code and handling "magic" modules differently), and
packages/api-utils/lib/self.js (which is now "magic" and provides a
"makeSelfModule" function).

A number of obsolete tests were removed. The async "define" statement
still works. E10S support is almost certainly broken.