Merge the two competing test-self.js (in api-utils and test-harness).
It turns out that the api-utils one fails when invoked under 'cfx testall',
because it was checking 'self.name' against the package name (api-utils), and
the implementation of self.name can't maintain that behavior under 'cfx
testall' as it can when run one-package-at-a-time under 'cfx test'.

But! It turns out that limitations in our require() syntax have caused
problems for colliding test-file-names under 'cfx testall' since forever (at
least 1.0b3), and the symptom was that the test-harness version of
test-self.js was run twice, while the api-utils version was not run. So we
never noticed, until Irakli's runtime-search-logic patch from today changed
the order, causing the api-utils version to be run twice and the test-harness
version to not be run.

Fixing the require() syntax is out-of-bounds for 1.0. The problem is that
unit-test-finder.js gets a directory name and a filename, and needs to
construct an argument for require(ARG) that will get the right module. Our
current search code doesn't know about "sections" like "test" (in which all
the tests live), so there's no ARG that can be passed which will point to the
right file. The longer term fix is to get rid of sections, and allow the test
finder to do something like require("addon-kit/tests/test-panel").

Also, it turns out that self.name was an undocumented property added
recently (1.0b5), kinda by accident. I plan to remove it from self.js, so
this patch also removes the assertion that looks at it, removing the
occasional failure from the api-utils version.