Bug 592512 - don't use constant filename for resultsfile. r=irakli
Previously, 'cfx run' and 'cfx test' on unix put results in a constant file
in /tmp/, so two users on the same box could not run simultaneously, and a
symlink attack was possible. OS-X uses a per-user tempfile, preventing the
symlink attack, but still preventing one user from running two sets of tests
at the same time.

This changes those commands to use a unique filename for each run, avoiding
both problems. It uses tempfile.mkstemp() to be python2.5-safe.