First cut at producing smaller "stripped" XPIs, including just the modules we actually use.
I'm not entirely confident about the accuracy, so this is opt-in: you must
use a --strip-xpi flag to activate the filter, otherwise you get the same
everything-under-the-sun-included XPIs as before. module-development/xpi.md
mentions this. Once this feature is seen to work well, we should make it the
default.

The XPI needs to include all "loader modules", which are modules that the
loader uses but which aren't referenced by normal code. They must not be
stripped, as otherwise the XPI will be broken and unusable.

In this version, the list of loader modules is hardcoded, and ought to be
built by running the manifest scanner over the known entry
points (cuddlefish.js, securable-module.js).

Also improve error messages in securable-module.js when lookup fails.