Bug 661615: update Translator tutorial for new linker. r=wbamberg,a=blocker
This changes the require() statements to avoid a collision between the name
of the package ("translator") and the name of a module in that
package (translator.js), switching to an explicit fully-qualified reference.
It also changes the name of the module to translate.js, to steer developers
away from further confusion down the road.