Bug 661929: don't append @jetpack to fully-qualified IDs. r=irakli,a=blocker
Now that we're no longer deriving addon-ids from ECDSA keys, it'd be nice to
use the "id" field in package.json as the addon-id directly, rather than
appending "@jetpack" to it first. This would also make it possible to build
addons with arbitrary email-address-style IDs, and would help convert
old-style XUL addons to the jetpack source format.

It would also help with bug 661376, which wants to round-trip addons between
the SDK and Builder, where the ID is read from install.rdf (extracted out of
the XPI bundle) and then stashed in package.json . The unconditional
appending of @jetpack led to an ever-growing ID string, one extra suffix per
roundtrip.

This patch changes the SDK to not append @jetpack when the "id" field in
package.json already has an "@" in it. It also makes sure that other strings
derived from the ID (such as resource:// URIs) can tolerate such IDs.

It does not yet tolerate {GUID}-style IDs, as the curly braces wind up in
some fields that cannot tolerate them. This will be fixed in a later release.

Internally, a new value ("uriPrefix") is computed by the SDK and passed to
the runtime (via harness-options.json), so the remaining runtime search
code (used by unit tests and specialized dynamic loaders) gets the right
value.

The "reading-data" example was changed to use a user@host -style ID, for
testing.