Bug 662471: cfx init should use lowercase name in uppercase directory. r+a=myk
When invoking 'cfx init' in an empty directory, the name of the directory is
used to set the addon's top-level package name. Package names must be
lowercase, and the directory name might have had uppercase characters, so
this patch downcases the string before setting the 'name' property in the
generated package.json.