remove api-utils/tests/test-securable-module, update test-file to match
Irakli tells me the functionality tested by this module is no longer relevant
to his new loader, so rather than try to rewrite it to avoid doing a runtime
search, I'm just getting rid of it.

test-file.js is still relevant (it exercises the "file" module), but it was
using the old interoperablejs-read-only/ directory as a source of known
filenames. I changed it to use modules/ for that purpose instead.