Bug 717409: allow install_addon() of pre-unpacked code, for mobile. r=mhammond
This fixes a regression that was introduced in bug 638742, because I didn't
realize that any cfx code was passing a *directory* into install_addon(). The
fennec code, during 'cfx run', uses this feature to install a special
"mobile-killer" addon (to help terminate the brower being tested), which
lives in the SDK itself. It'd be inconvenient to either change 'cfx run' to
build mobile-killer.xpi each time, or to commit mobile-killer.xpi into the
source tree (eww). So this fix lets us continue to use it in-place.

Closes bug 717409.