Bug 721257: get version from parentdir when .git is unavailable. r=dcm
This updates the embedded _version.py to upstream versioneer-0.6, which adds
code that can get the version number from the name of the parent directory.
For example, if the SDK finds itself living in a directory named
"addon-sdk-1.4.1", then (if it doesn't find variables expanded by
git-archive, and doesn't find a usable .git directory for git-describe) it
will strip the "addon-sdk-" prefix and describe itself as "1.4.1".

This should address the needs of Flightdeck (aka Addon Builder), which uses
git submodules to obtain a copy of the SDK, but then strips the .git
directories when deploying to the frontend web servers. This leaves neither
expanded variables nor .git metadata, but *does* use the addon-sdk-N.N/
directory name convention.