Bug 1141275: Fix l10n/html in jpm.
This reverts 83baec9d671b0fba173525e5c4e97b033dacc69d which had the
unintentional side-effect of breaking l10n/html in jpm. It also makes the code
that loads locales in test/addons/l10n handle the missing json file in the same
was as sdk/addon/runner.

To get the l10n test add-on working in jpm tests I've moved the en-GB locale to
en-US. This is because of bug 1141696. l10n-properties is still disabled but
that is essentially testing the jpm behaviour in cfx so we shouldn't need to
worry about it.

Switched from pageworker to a tab in the test add-on so in e10s mode the tab
will load remotely.