Bug 1132753: Fix sdk/l10n/html to not use CPOWs and shims. r=erikvold
This reverts 83baec9d671b0fba173525e5c4e97b033dacc69d which had the
unintentional side-effect of breaking l10n/html in jpm. It also makes the code
that loads locales in test/addons/l10n handle the missing json file in the same
way as sdk/addon/runner.

To get the l10n test add-on working in jpm tests I've moved the en-GB locale to
en-US. This is because of bug 1141696. l10n-properties is still disabled but
that is essentially testing the jpm behaviour in cfx so we shouldn't need to
worry about it.

Splits off the document listening code to run in content processes allowing us
to avoid CPOWs.

Added a test that pages loaded in tabs are localized and copied the entire
l10n add-on to an e10s test add-on so we verify that remote pages are localized
too.