Bug 1146926: Allow sending CPOWs from parent to child processes. r=gabor
Defines an emitCPOW(event, arguments, cpows) to send a dictionary of CPOW
objects to a child process or frame. The only process that will receive the
event is the one that owns all of the CPOWs sent. The CPOW dictionary will be
appended to the arguments emitted from the port.

Also gets rid of some of the redundancy in message handling so we can more
easily change the protocol in the future.